see #37: figured out memory problem, not sure how to fix.
Dynamic tests are all queued up (if you run the suite right now you will see the comparisons (aka the actual run of the executor method in the dynamic test case) all running at the end of the suite. _BUT_: these take a Yresult, which contains a Ychange, which contains a RepositoryService which kindly maintains a large cache. This means that the StartEnvironments aren't being cleared out as we go through the tests as they all need to stick around until the end.

This is also why the tests labels don't update as the tests are made; junit doesn't know about the contents of the test array until it's fully built (which is at the end of the whole process).