see #47 make it so envnames and skipnames behave the same (aka can take a list of includes/skips). rename ENV_NAME to ENV_NAMES. Simplify the include/exclude logic in maindynamic test.
More controversially: removed the caching size checks from the CachingRepositoryService. Right now this doesn't seem to matter (runs the whole test suite fine). Might have to revisit if it's a problem in the deployed docker container. Mainly motivated to do this due to the JVM warnings about carrotsearch.sizeof.RamUsageEstimator on every execution.